Change API for removing and detecting entities
Before this commit, there was a function get_component_of_type() which took a component type and also a get_component() method which took a component instance. Now, there is only a get_component() method which takes the component *type*. The reason for this is that it is likely that whenever the component instance is available, its type will also be, so having both methods is redundant.