Make systems objects instead of functions
All system functions must have an update method or else an AssertionError will be raised. The reason for this change is to allow systems to store data more easily, as this can be done as attributes of the system object and then modified by update. This paradigm will support implementation of outboard ECS, where objects are created and stored by the system in a dictionary, with keys for each entity.